requirements.txt
pytorch-lightning
hydra-core
wandb